+++ package.json
@@ -14,10 +14,8 @@
   "dependencies": {
     "@types/three": "^0.168.0",
     "gh-pages": "^6.1.1",
-    "gsap": "^3.12.5",
     "react": "^18.3.1",
     "react-dom": "^18.3.1",
-    "react-multi-carousel": "^2.8.5",
     "react-router-dom": "^6.26.2",
     "three": "^0.168.0",
     "zustand": "^5.0.0-rc.2"
