+++ src/components/DDButton.tsx
@@ -1,9 +1,9 @@
 import { useRef } from "react";
-import ButtonAudio from '../assets/sounds/navigation-digital-menu-click.wav'
+import ButtonAudio from '../assets/sounds/button_sound.mp3'
 import { useNavigate } from "react-router-dom";
 
 const DDButton = ({ href, onClick, className, children,  }) => {
-    const audioRef = useRef(null);
+    const audioRef = useRef(null); // Referencia para el audio
     const navigate = useNavigate();
     const handleClick = () => {
         if(onClick){
