--- src/components/Footbar.js
@@ -1,1 +1,0 @@
-export {};
