--- src/hooks/useFetchData.js
@@ -1,27 +1,0 @@
-import { useState, useEffect } from 'react';
-const useFetchJson = (url) => {
-    const [data, setData] = useState(null);
-    const [loading, setLoading] = useState(true);
-    const [error, setError] = useState(null);
-    useEffect(() => {
-        const fetchData = async () => {
-            try {
-                const response = await fetch(url);
-                if (!response.ok) {
-                    throw new Error(`HTTP error! status: ${response.status}`);
-                }
-                const json = await response.json();
-                setData(json);
-            }
-            catch (err) {
-                setError(err);
-            }
-            finally {
-                setLoading(false);
-            }
-        };
-        fetchData();
-    }, [url]);
-    return { data, loading, error };
-};
-export default useFetchJson;
