--- src/main.js
@@ -1,6 +1,0 @@
-import { jsx as _jsx } from "react/jsx-runtime";
-import { StrictMode } from 'react';
-import { createRoot } from 'react-dom/client';
-import App from './App.tsx';
-import './index.css';
-createRoot(document.getElementById('root')).render(_jsx(StrictMode, { children: _jsx(App, {}) }));
