--- src/pages/02.TravelCustomization/CardPlanetarySystem.css
@@ -1,118 +1,0 @@
-.text-title{
-    display: flex;
-    justify-content: center;
-    padding: 1.2rem;
-    font-size: 1.5rem;
-}
-
-
-.image-card {
-    display: flex;
-    flex-direction: column;
-    justify-content: end;
-    height: 80vh;
-    background-size: cover;
-    background-position: center;
-    border: 2px solid white;
-}
-
-.ps-1{
-    background-image: url('../../assets/planetary-systems/sistema-planetario-1.jpg');
-}
-.ps-2{
-    background-image: url('../../assets/planetary-systems/sistema-planetario-2.jpg');
-}
-.ps-3{
-    background-image: url('../../assets/planetary-systems/sistema-planetario-3.jpg');
-}
-.ps-4{
-    background-image: url('../../assets/planetary-systems/sistema-planetario-4.jpg');
-}
-
-.image-card:hover {
-    transform: scale(1.05);
-    box-shadow: 0 0 15px 4px rgb(255, 145, 19);
-    cursor: pointer;
-}
-
-.list-images {
-    display: grid;
-    grid-template-columns: repeat(4, 1fr);
-    gap: 3rem;
-    margin: 0rem 6rem
-}
-
-.container {
-    background-image: url('../../assets/background-selection.jpg');
-    background-size: cover;
-    background-position: center;
-    min-width: 100vw;
-    min-height: 100vh;
-}
-
-/* .sky {
-    min-width: 100vw;
-    min-height: 100vh;
-    position: relative;
-    background: black;
-    overflow: hidden;
-    display: flex;
-    justify-content: center;
-    align-items: center;
-  }
-  
-
-.star {
-    position: absolute;
-    background-color: rgba(255, 255, 255, 0.8);
-    opacity: 0;
-    animation: twinkle 5s infinite ease-in-out;
-      box-shadow: 0px 0px 15px 4px rgba(255, 255, 255, 0.8);
-    width: 200px;  
-    aspect-ratio: 1;
-    background: #faf1ca;
-    clip-path: polygon(50% 0,
-      calc(50%*(1 + sin(.4turn))) calc(50%*(1 - cos(.4turn))),
-      calc(50%*(1 - sin(.2turn))) calc(50%*(1 - cos(.2turn))),
-      calc(50%*(1 + sin(.2turn))) calc(50%*(1 - cos(.2turn))),
-      calc(50%*(1 - sin(.4turn))) calc(50%*(1 - cos(.4turn))) 
-     ); 
-  }
-  
-  .sky .star:nth-child(1) { width: 2px; height: 2px; top: 10%; left: 15%; animation-delay: 0s; }
-  .sky .star:nth-child(2) { width: 3px; height: 3px; top: 25%; left: 30%; animation-delay: 1.5s; }
-  .sky .star:nth-child(3) { width: 4px; height: 4px; top: 40%; left: 20%; animation-delay: 2.5s; }
-  .sky .star:nth-child(4) { width: 6px; height: 6px; top: 50%; left: 60%; animation-delay: 3s; }
-  .sky .star:nth-child(5) { width: 3px; height: 3px; top: 70%; left: 80%; animation-delay: 0.5s; }
-  .sky .star:nth-child(6) { width: 5px; height: 5px; top: 15%; left: 75%; animation-delay: 2s; }
-  .sky .star:nth-child(7) { width: 2px; height: 2px; top: 60%; left: 30%; animation-delay: 1s; }
-  .sky .star:nth-child(8) { width: 6px; height: 6px; top: 85%; left: 45%; animation-delay: 3.5s; }
-  .sky .star:nth-child(9) { width: 4px; height: 4px; top: 5%; left: 55%; animation-delay: 2s; }
-  .sky .star:nth-child(10) { width: 3px; height: 3px; top: 75%; left: 25%; animation-delay: 4s; }
-  .sky .star:nth-child(11) { width: 5px; height: 5px; top: 65%; left: 10%; animation-delay: 0.3s; }
-  .sky .star:nth-child(12) { width: 2px; height: 2px; top: 45%; left: 90%; animation-delay: 1.7s; }
-  .sky .star:nth-child(12) { width: 2px; height: 2px; top: 45%; left: 90%; animation-delay: 1.7s; }
-  
-
-  @keyframes twinkle {
-    0% {
-      opacity: 0;
-      transform: scale(0.5);
-      box-shadow: 0px 0px 5px 1px rgba(255, 255, 255, 0.3), 0px 0px 8px 2px rgba(255, 255, 255, 0.2);
-    }
-    50% {
-      opacity: 1;
-      transform: scale(1.5);
-      box-shadow: 0px 0px 20px 5px rgba(255, 255, 255, 1), 0px 0px 25px 6px rgba(255, 255, 255, 0.8);
-    }
-    100% {
-      opacity: 0;
-      transform: scale(0.5);
-      box-shadow: 0px 0px 5px 1px rgba(255, 255, 255, 0.3), 0px 0px 8px 2px rgba(255, 255, 255, 0.2);
-    }
-  } */
-
-.text-overlay {
-    padding: 2rem 1.5rem;
-    color: white;
-}
