+++ src/pages/02.TravelCustomization/MisionDescription.css
@@ -10,22 +10,7 @@
 
 .astronaut {
     transform: scaleX(-1);
-    animation: float 4s ease-in-out infinite;
 }
-
-@keyframes float {
-    0% {
-      transform: translateY(0);
-    }
-    50% {
-      transform: translateY(30px);
-    }
-    100% {
-      transform: translateY(0);
-    }
-  }
-
-
 
 
 .left {
