+++ src/pages/02.TravelCustomization/SelectPlanetarySystem.tsx
@@ -1,18 +1,19 @@
 import { Link } from "react-router-dom";
-import CardPlanetarySystem from "./CardPlanetarySystem";
+import exoplanetsData from '../../data/exoplanets.json'
+import useMisionStore from "../../store/store";
+import DDButton from "../../components/DDButton";
 const SelectPlanetarySystem = () => {
-    
+    const { planetarySystem, setPlanetarySystem } = useMisionStore();
+
+    const changePlanetarySystem = ( option : string) => {
+        const filtered = exoplanetsData.filter((item) => item.solar_system === option)
+        setPlanetarySystem("Planetary System : " + option, filtered);
+    }
     return (
         <div className='container'>
-            <div className="text-title">
-                <p className="">Choose your destination</p>
-            </div>
-            <div className="list-images">
-                <CardPlanetarySystem text="Alfa Centauir" clas='ps-1'></CardPlanetarySystem>
-                <CardPlanetarySystem text="Tau Ceti" clas='ps-2'></CardPlanetarySystem>
-                <CardPlanetarySystem text="Gliese 1061" clas='ps-3'></CardPlanetarySystem>
-                <CardPlanetarySystem text="YZ Ceti" clas='ps-4'></CardPlanetarySystem>
-            </div>
+            <Link to="/personalization" style={{ textDecoration: 'none' }}><button className="button">Next</button></Link>
+            <DDButton href="/mapping" onClick={()=>changePlanetarySystem("Alfa Centauir")}>Sistema Planetario 1</DDButton>
+            <DDButton href="/mapping" onClick={()=>changePlanetarySystem("Alfa Centauir")}>Sistema Planetario 2</DDButton>
         </div>
     );
 }
