--- src/pages/02.TravelCustomization/SelectSpaceship.css
@@ -1,6 +1,0 @@
-.bottom {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  padding: 1rem 0rem;
-}
