--- src/pages/03.Mapping/Loading.js
@@ -1,5 +1,0 @@
-import { jsx as _jsx } from "react/jsx-runtime";
-const Loading = () => {
-    return (_jsx("div", { className: 'container' }));
-};
-export default Loading;
