+++ tsconfig.app.tsbuildinfo
@@ -1,1 +1,1 @@
-{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/custombackground.tsx","./src/components/ddbutton.tsx","./src/components/ddmodal.tsx","./src/components/ddtitle.tsx","./src/components/footbar.tsx","./src/components/navbar.tsx","./src/components/rotatingplanet.tsx","./src/hooks/useexoplanets.tsx","./src/hooks/usefetchdata.tsx","./src/pages/01.landingpage/landingpage.tsx","./src/pages/02.travelcustomization/cardplanetarysystem.tsx","./src/pages/02.travelcustomization/misiondescription.tsx","./src/pages/02.travelcustomization/selectplanetarysystem.tsx","./src/pages/02.travelcustomization/selectspaceship.tsx","./src/pages/03.mapping/loading.tsx","./src/pages/03.mapping/mapping.tsx","./src/pages/04.planetexploration/modalicon.tsx","./src/pages/04.planetexploration/planetinformation.tsx","./src/pages/04.planetexploration/video.tsx","./src/store/store.tsx"],"version":"5.6.2"}
+{"root":["./src/app.tsx","./src/main.tsx","./src/vite-env.d.ts","./src/components/ddbutton.tsx","./src/components/ddmodal.tsx","./src/components/ddtitle.tsx","./src/components/footbar.tsx","./src/components/navbar.tsx","./src/components/rotatingplanet.tsx","./src/hooks/useexoplanets.tsx","./src/hooks/usefetchdata.tsx","./src/pages/01.landingpage/landingpage.tsx","./src/pages/02.travelcustomization/misiondescription.tsx","./src/pages/02.travelcustomization/personalization.tsx","./src/pages/02.travelcustomization/selectplanetarysystem.tsx","./src/pages/03.mapping/loading.tsx","./src/pages/03.mapping/mapping.tsx","./src/pages/04.planetexploration/modalicon.tsx","./src/pages/04.planetexploration/planetinformation.tsx","./src/pages/04.planetexploration/video.tsx","./src/store/store.tsx"],"errors":true,"version":"5.6.2"}
