+++ tsconfig.node.json
@@ -7,18 +7,16 @@
 
     /* Bundler mode */
     "moduleResolution": "bundler",
+    "allowImportingTsExtensions": true,
     "isolatedModules": true,
     "moduleDetection": "force",
-    "noEmit": true, // Permitir la emisión
-    "allowImportingTsExtensions": true, // Añadir esta línea
-    
+    "noEmit": true,
+
     /* Linting */
-    "strict": false,  // Relajar las reglas
-    "noUnusedLocals": false,
-    "noUnusedParameters": false,
-    "noFallthroughCasesInSwitch": true,
-    "noImplicitAny": false,
-    "noEmitOnError": false
+    "strict": true,
+    "noUnusedLocals": true,
+    "noUnusedParameters": true,
+    "noFallthroughCasesInSwitch": true
   },
   "include": ["vite.config.ts"]
 }
